#4 - segmented control 생성
segmented control 생성하고 밑에 underline rounding한 거 triangle로 바꾸는거 고민 중
